Add the rover12_description package.
Added a URDF to describe the coordinate frames of our robot, with some
visual models for the body and wheels. The rear and front wheels can
rotate and the front wheels can steer.

This also adds some useful launch files for visualizing the URDF and
loading the robot configuration (launch/rover12.launch).

The configuration of the joint_state_publisher and
robot_state_publisher have been broken out into YAML files in the
config directory. There are 6 topics that joint_state_publisher is
subscribing to, one for each of the degrees of freedom of the URDF. All
these topics are under the /joints/ namespace.